5.25饭前一存

+++ App/Inc/yl_searchline.h
@@ -27,8 +27,8 @@ void Goin_rightcircul();
 void Goin_leftcircul();
 int8 Count_black(int16 jh, int8 start, int8 end, int8 extent);
 
-void FullLine(Site_t, Site_t, int8 line[]);
-uint8 IsStraight(int8 line[]);
+void FullLine(Site_t, Site_t, int16 line[]);
+uint8 IsStraight(int16 line[]);
 void FullBend(int8, int8, int8);
 
 #endif
